+++ lib/htmlwhitelist.dart
@@ -2,6 +2,7 @@
 // All rights reserved. Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
+export 'package:htmlwhitelist/src/api/attributecollector.dart';
 export 'package:htmlwhitelist/src/api/cleaner.dart';
 export 'package:htmlwhitelist/src/api/typedefs.dart';
 export 'package:htmlwhitelist/src/api/whitelist.dart';