+++ lib/src/api/typedefs.dart
@@ -2,6 +2,8 @@
 // All rights reserved. Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
+import 'package:htmlwhitelist/htmlwhitelist.dart';
+
 /// Returns `true` if the given [tag] and [attributes] are accepted.
 ///
 /// [attributes] contains all attribute names and values of the source tag.
@@ -10,16 +12,13 @@ typedef bool Filter(String tag, Map<String, String> attributes);
 /// Returns `true` if the given [name] matches.
 typedef bool Matcher(String name);
 
-/// Invokes the [adder] to generate attributes for the given [tag].
+/// Uses the [collector] to generate attributes for the given [tag].
 ///
 /// [originalAttributes] contains all attribute names and values in the source
 /// tag. This makes it possible to create attributes and values based on
 /// the original tag.
-typedef void AttributeGenerator(
-    String tag, Map<String, String> originalAttributes, AddAttribute adder);
-
-/// Adds an attribute with the given [name] and [value] to the current tag.
-typedef void AddAttribute(String name, String value);
+typedef void AttributeGenerator(String tag,
+    Map<String, String> originalAttributes, AttributeCollector collector);
 
 /// Matches all tags.
 final Matcher anyTag = (t) => true;
@@ -37,6 +36,6 @@ final Filter always = (t, a) => true;
 ///
 ///     Whitelist.none
 ///         .tags('a')
-///         .extraAttributes('a', forceAttribute('target', '_blank'));
-AttributeGenerator forceAttribute(String name, String value) =>
-    ((t, o, a) => a(name, value));
+///         .extraAttributes('a', setAttribute('target', '_blank'));
+AttributeGenerator setAttribute(String name, String value) =>
+    ((t, o, c) => c[name] = value);